Swapping out the useFeaturePoint
